Move TailoredSearch class to its own file
